requirements.txt
python-dotenv
langchain-groq
pandas
numpy==1.25.2
pandasai==1.5.15

